add sketch Statictics logic #2

+++ src/main/java/ru/javarush/ogarkov/island/factory/Factory.java
@@ -5,7 +5,7 @@ import ru.javarush.ogarkov.island.settings.Items;
 
 public interface Factory{
     BasicItem createItem(Items item);
-    int getCreatedItemsCount();
+    long getCreatedItemsCount();
